covidgraphs/requirements.txt
asgiref==3.4.1
Django==3.2.6
pip==21.1.2
python-dateutil==2.8.2
pytz==2021.1
setuptools==57.0.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
wheel==0.36.2
